Features

The core part of FLOYao is the a new register type, the MajoranaReg:

FLOYao.MajoranaReg — Type

MajoranaReg{T} <: AbstractRegister{2}
MajoranaReg(state::AbstractMatrix{T<:Real})

A register holding the "state" of a Majorana operators when propagating through a FLO circuit as a 2n×2n matrix.

Warning

The MajoranaReg constructor will not initialize the state matrix. It is recommended to use FLOYao.zero_state or FLOYao.product_state to produce your initial state.

State initialization

There are several functions provided, to create MajoranaReg's in different states. They are not exported from FLOYao.jl in order to avoid name collisions with Yao.jl

FLOYao.zero_state — Function

zero_state([T=Float64,] n)

Create a Majorana register on n qubits in the vacuum state $|Ω⟩$ with storage type T.

FLOYao.zero_state_like — Function

zero_state_like(reg::AbstractRegister)

Create a Majorana register in the zero state with the same element type and number of qubits as reg.

FLOYao.one_state — Function

one_state([T=Float64,] n)

Create a Majorana register on n qubits in the all one state $|1 ⋯ 1⟩$ with storage type T.

FLOYao.product_state — Function

product_state([T=Float64,] bit_str::DitStr{2})
product_state([T=Float64,] bit_configs::AbstractVector)
product_state([T=Float64,] nbits::Int, val::Int)

Create an MajoranaReg of a product state.

The state can be specified as a bit string, as an array of Integers or Booleans or with nbits and val.

FLOYao.rand_state — Function

rand_state([Float64,] n)

Create a Haar random MajoranaReg on n qubits.

as well as some functions to reset MajoranaReg's to fixed states:

FLOYao.zero_state! — Function

zero_state!(reg::MajoranaReg)

Put reg into the computational zero state

FLOYao.one_state! — Function

one_state!(reg::MajoranaReg)

Put reg into the all ones state

FLOYao.product_state! — Function

product_state!(reg::MajoranaReg, bit_str::BitStr)

Put reg into the product state described by bit_str

Applying gates

Application of Supported gates to MajoranaReg's works as it does in Yao.jl

reg = FLOYao.zero_state(2)
gate = put(2, 1 => X)

# using apply!
apply!(reg, gate)

# or using the pipe syntax
reg |> gate

# or the non-mutating version
result = apply(reg, gate)

Measuring expectation values

Measuring expectation values with respect to (almost) arbitrary Yao.AbstractBlocks works exactly the same way it does in Yao.jl using expect:

YaoAPI.expect — Function

expect(op::Union{AbstractBlock,MajoranaSum}, reg::MajoranaReg)
expect(op::Union{AbstractBlock,MajoranaSum}, (reg => circuit)::Pair{<:MajoranaReg,<:AbstractBlock})

Get the expectation value of an operator, the second parameter can be a register reg or a pair of input register and circuit reg => circuit.

expect'(op::Union{AbstractBlock,MajoranaSum}, reg => circuit::) -> Pair{<:MajoranaReg,<:AbstractVector}
expect'(op::Union{AbstractBlock,MajoranaSum}, reg::MajoranaReg) -> MajoranaReg

Obtain the gradient with respect to registers and circuit parameters. For pair input, the second return value is a pair of gψ => gparams, with gψ the gradient of input state and gparams the gradients of circuit parameters. For register input, the return value is a register.

But if you need convert expectation values many times with respect to the same observable op it is recommended to convert the Yao.AbstractBlock to a MajoranaSum which acts as a (more performant) replacement of op in expect and expect':

FLOYao.MajoranaSum — Type

MajoranaSum{T}
MajoranaSum(terms::Vector{MajoranaTerm{T}})
MajoranaSum(terms::Dict{Vector{Int}, T})
MajoranaSum(op::AbstractBlock)

A sum of FLOYao.MajoranaTerms to be used as a more performant replacement of op::Yao.AbstractBlock in expect and expect'

Yao.AbstractBlocks that are sums of (scaled) tensor products of Pauli gates can be directly converted to MajoranaSums:

julia> yaoham = (put(4, 1=>Z) + 2kron(4, 1=>X, 2=>Z, 3=>Z, 4=>X) + 3.5put(4, 2=>Z)
                 + 0.5kron(4, 1=>Z, 2=>Z) - kron(4, 2 => X, 4 => Y));

julia> MajoranaSum(yaoham)
5-element MajoranaSum{ComplexF64}:
  [4, 5, 6, 8] : -1.0 - 0.0im
  [2, 7] : 0.0 + 2.0im
  [1, 2, 3, 4] : -0.5 - 0.0im
  [1, 2] : 0.0 - 1.0im
  [3, 4] : 0.0 - 3.5im

FLOYao.MajoranaTerm — Type

MajoranaTerm{T}
MajoranaTerm(coeff::Number, indices::Vector{Int})

A Majorana monomial represented by its coefficient and indices on which it acts non-trivially.

Sampling

Samples in the computational basis can be obtained with the same functions as in Yao.jl:

YaoAPI.measure — Method

measure(reg::MajoranaReg[, locs]; nshots=1, rng=Random.GLOBAL_RNG) -> Vector{BitStr}

Measure a MajoranaReg and return the results as a vector of BitStrings. This is a cheating version of measure! that does not need to recompute the reg for each new sample and also does not alter it.

Arguments

reg::MajoranaReg: The state register

locs: The list of qubits to measure. Defaults to all qubits.

nshots=1: The number of samples to take

rng=Random.GLOBAL_RNG: The random number generator to use

YaoAPI.measure! — Method

measure!([postprocess::Union{NoPostProcess, ResetTo},] reg::MajoranaReg; rng=Random.GLOBAL_RNG)

Measure a Majorana register in the computational basis and return the resulting BitStr.

Arguments

postprocess: Is the post processing method

NoPostProcess() (the default) will collapse the state to the measurement outcome state.

ResetTo(bit_str) will reset the state to the product state specified by bit_str

reg::MajoranaReg: The quantum register

rng::Random.GLOBAL_RNG: The RNG to use

Fidelities and bitstring probabilities

Fidelities between two MajoranaRegs can be computed the same way as in Yao.jl:

YaoAPI.fidelity — Method

fidelity(reg1::MajoranaReg, reg2::MajoranaReg)
fidelity'(pair_or_majoranareg, pair_or_majoranareg) -> (g1, g2)

The fidelity $|⟨ψ|φ⟩|$ between the FLO states $|ψ⟩$ and $|φ⟩$ defined by reg1 and reg2.

The gradient with respect to registers and circuit parameters. For a pair input reg => circuit the returned gradient is a pair of greg => gparams with greg the gradient with respect to the input state and gparams the gradient with respect to circuit parameters. For a reg input the return value is the gradient with respect to the register.

Note

This definition differs from the standard definition $|⟨φ|ψ⟩|²$ by a square root, but is consistent with the one used in Yao.jl

Under the hood this function uses a non-exported function to compute the probability of obtaining a given bit_string as an output when measuring in the computational basis:

FLOYao.bitstring_probability — Function

bitstring_probability(reg::MajoranaReg, bit_string::BitStr)

The probability to measure a given bit_string when measuring reg

Todo

Rewrite this using pfaffians

Non-exported functions

The following functions are not exported and not really needed for the functionality of FLOYao, but can be useful for debugging.

FLOYao.paulibasis2qubitop — Function

paulibasis2qubitop(P::AbstractVector)

Converts an operator to in the pauli basis to a matrix in the computational basis. Inverse to qubit2paulibasis.

FLOYao.majoranasquares2qubitbasis — Function

majoranasquares2qubitbasis(H::AbstractMatrix)

Converts an $2n×2n$ Majorana hamiltonian H into the full $2^n×2^n$ hamiltonian in the qubit basis.

FLOYao.qubit2paulibasis — Function

qubit2paulibasis(A::AbstractMatrix)

Converts a $2^n×2^n$ matrix A in the standard qubit basis into a $4^n$ vector representing the same operator in the Pauli basis.

The ordering is as follows: Let $σ^0 = I, σ^1 = X, σ^2 = Y$ and $σ^3 = Z$.

Todo

Creating the dense pauli tensor product via kron and inner product via dot is much slower than neccessary, since the pauli tensor product matrix is very sparse. Much faster would be to compute the inner product directly.

FLOYao.qubit2majoranaevolution — Function

qubit2majoranaevolution(U::AbstractMatrix, locs)

Turns a $n$ qubit unitary U on the $n$ qubits in locs into the corresponding $SO(2n×2n)$ matrix for the evolution of the Majorana operators.

FLOYao.paulibasis2majoranasquares — Function

paulibasis2majoranasquares(P::AbstractVector, locs=1:log4(length(P)))

Convert an operator written in the Pauli basis as a $4^n$-element vector to the corresponding $2n×2n$ matrix of coefficients of products of two Majorana operators.

Throws a NonFLOException if P contains terms that are not corresponding to the product of two Majorana operators.

FLOYao.yaoham2majoranasquares — Function

yaoham2majoranasquares(::Type{T}=Float64, yaoham::AbstracBlock{2})

Convert a hamiltonian written as a YaoBlock into the corresponding $2n×2n$ majorana hamiltonian.

FLOYao.yaoblock2majoranasum — Function

yaoham2majoranasum(::Type{T}=Float64, yaoham::AbstracBlock{2})

Convert a YaoBlock into a MajoranaSum{complex(T)}

FLOYao.majorana2arrayreg — Function

majorana2arrayreg(reg::MajoranaReg)

Converts a 2n×2n MajoranaReg reg into a 2^n ArrayReg.

Note

This implementation is not very clever and should mainly be used for debugging purposes with small numbers of qubits. It pipes a random state $|ψ⟩$ through the projector $U|Ω⟩⟨Ω|U^†$ which may give inaccurate results if $⟨ψ|U|ψ⟩$ is very small.

FLOYao.random_orthogonal_matrix — Function

random_orthogonal_matrix([::Type{T}=Float64,] n)

Generate a Haar random matrix in $O(n)$ with element type T using the algorithm described in How to generate random matrices from the classical compact groups