Features

The core part of FLOYao is the a new register type, the MajoranaReg:

FLOYao.MajoranaReg — Type

MajoranaReg{T} <: AbstractRegister{2}
MajoranaReg(state::AbstractMatrix{T<:Real})

A register holding the "state" of a Majorana operators when propagating through a FLO circuit as a 2n×2n matrix.

Warning

The MajoranaReg constructor will not initialize the state matrix. It is recommended to use FLOYao.zero_state or FLOYao.product_state to produce your initial state.

State initialization

There are several functions provided, to create MajoranaReg's in different states. They are not exported from FLOYao.jl in order to avoid name collisions with Yao.jl

FLOYao.zero_state — Function

zero_state([T=Float64,] n)

Create a Majorana register on n qubits in the vacuum state $|Ω⟩$ with storage type T.

FLOYao.zero_state_like — Function

zero_state_like(reg::AbstractRegister)

Create a Majorana register in the zero state with the same element type and number of qubits as reg.

FLOYao.one_state — Function

one_state([T=Float64,] n)

Create a Majorana register on n qubits in the all one state $|1 ⋯ 1⟩$ with storage type T.

FLOYao.product_state — Function

product_state([T=Float64,] bit_str::DitStr{2})
product_state([T=Float64,] bit_configs::AbstractVector)
product_state([T=Float64,] nbits::Int, val::Int)

Create an MajoranaReg of a product state.

The state can be specified as a bit string, as an array of Integers or Booleans or with nbits and val.

as well as some functions to reset MajoranaReg's to fixed states:

FLOYao.zero_state! — Function

zero_state!(reg::MajoranaReg)

Put reg into the computational zero state

FLOYao.one_state! — Function

one_state!(reg::MajoranaReg)

Put reg into the all ones state

FLOYao.product_state! — Function

product_state!(reg::MajoranaReg, bit_str::BitStr)

Put reg into the product state described by bit_str

Applying gates

Application of Supported gates to MajoranaReg's works as it does in Yao.jl

YaoAPI.apply! — Function

apply!(register, block)

Apply a block (of quantum circuit) to a quantum register.

Note

to overload apply! for a new block, please overload the unsafe_apply! function with same interface. Then the apply! interface will do the size checks on inputs automatically.

Examples

julia> r = zero_state(2)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 2/2
    nlevel: 2

julia> apply!(r, put(2, 1=>X))
ArrayReg{2, ComplexF64, Array...}
    active qubits: 2/2
    nlevel: 2

julia> measure(r;nshots=10)
10-element Vector{DitStr{2, 2, Int64}}:
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎
 01 ₍₂₎

Measuring expectation values

The same goes measuring expectation values

YaoAPI.expect — Function

expect(op::AbstractBlock, reg) -> Vector
expect(op::AbstractBlock, reg => circuit) -> Vector
expect(op::AbstractBlock, density_matrix) -> Vector

Get the expectation value of an operator, the second parameter can be a register reg or a pair of input register and circuit reg => circuit.

expect'(op::AbstractBlock, reg=>circuit) -> Pair
expect'(op::AbstractBlock, reg) -> AbstracRegister

Obtain the gradient with respect to registers and circuit parameters. For pair input, the second return value is a pair of gψ=>gparams, with gψ the gradient of input state and gparams the gradients of circuit parameters. For register input, the return value is a register.

Note

For batched register, expect(op, reg=>circuit) returns a vector of size number of batch as output. However, one can not differentiate over a vector loss, so expect'(op, reg=>circuit) accumulates the gradient over batch, rather than returning a batched gradient of parameters.

Examples

julia> r = normalize!(product_state(bit"11") + product_state(bit"00"))
ArrayReg{2, ComplexF64, Array...}
    active qubits: 2/2
    nlevel: 2

julia> op = chain(2, put(1=>H), put(2=>X))
nqubits: 2
chain
├─ put on (1)
│  └─ H
└─ put on (2)
   └─ X

julia> expect(op, r)
0.7071067811865474 + 0.0im

where (for now) op is restricted to be an observable quadratic in the Majorana operators (see Known restrictions).

Sampling

Samples in the computational basis can be obtained with the same functions as in Yao.jl:

YaoAPI.measure — Method

measure(reg::MajoranaReg[, locs]; nshots=1, rng=Random.GLOBAL_RNG) -> Vector{BitStr}

Measure a MajoranaReg and return the results as a vector of BitStrings. This is a cheating version of measure! that does not need to recompute the reg for each new sample and also does not alter it.

Arguments

reg::MajoranaReg: The state register

locs: The list of qubits to measure. Defaults to all qubits.

nshots=1: The number of samples to take

rng=Random.GLOBAL_RNG: The random number generator to use

YaoAPI.measure! — Method

measure!([postprocess::Union{NoPostProcess, ResetTo},] reg::MajoranaReg; rng=Random.GLOBAL_RNG)

Measure a Majorana register in the computational basis and return the resulting BitStr.

Arguments

postprocess: Is the post processing method

NoPostProcess() (the default) will collapse the state to the measurement outcome state.

ResetTo(bit_str) will reset the state to the product state specified by bit_str

reg::MajoranaReg: The quantum register

rng::Random.GLOBAL_RNG: The RNG to use

Non-exported functions

The following functions are not exported and not really needed for the functionality of FLOYao, but can be useful for debugging.

FLOYao.paulibasis2qubitop — Function

paulibasis2qubitop(P::AbstractVector)

Converts an operator to in the pauli basis to a matrix in the computational basis. Inverse to qubit2paulibasis.

FLOYao.majoranasquares2qubitbasis — Function

majoranasquares2qubitbasis(H::AbstractMatrix)

Converts an $2n×2n$ Majorana hamiltonian H into the full $2^n×2^n$ hamiltonian in the qubit basis.

FLOYao.qubit2paulibasis — Function

qubit2paulibasis(A::AbstractMatrix)

Converts a $2^n×2^n$ matrix A in the standard qubit basis into a $4^n$ vector representing the same operator in the Pauli basis.

The ordering is as follows: Let $σ^0 = I, σ^1 = X, σ^2 = Y$ and $σ^3 = Z$.

Todo

Creating the dense pauli tensor product via kron and inner product via dot is much slower than neccessary, since the pauli tensor product matrix is very sparse. Much faster would be to compute the inner product directly.

FLOYao.qubit2majoranaevolution — Function

qubit2majoranaevolution(U::AbstractMatrix, locs)

Turns a $n$ qubit unitary U on the $n$ qubits in locs into the corresponding $SO(2n×2n)$ matrix for the evolution of the Majorana operators.

FLOYao.paulibasis2majoranasquares — Function

paulibasis2majoranasquares(P::AbstractVector, locs=1:log4(length(P)))

Convert an operator written in the Pauli basis as a $4^n$-element vector to the corresponding $2n×2n$ matrix of coefficients of products of two Majorana operators.

Throws a NonFLOException if P contains terms that are not corresponding to the product of two Majorana operators.

FLOYao.yaoham2majoranasquares — Function

yaoham2majoranasquares(::Type{T}=Float64, yaoham::AbstracBlock{2})

Convert a hamiltonian written as a YaoBlock into the corresponding $2n×2n$ majorana hamiltonian.

FLOYao.majorana2arrayreg — Function

majorana2arrayreg(reg::MajoranaReg)

Converts a 2n×2n MajoranaReg reg into a 2^n ArrayReg.

Note

This implementation is not very clever and should mainly be used for debugging purposes with small numbers of qubits. If $⟨+⋯ +|U|+⋯ +⟩$ is close to zero, it is not very accurate.